Use ActiveModel validations in form object.
Avoids the need to create our own valid? method, and also means we can run validations before saving anything to the database. With the previous code, if you had a valid volunteer but an invalid shift, the volunteer would still be saved to the database when calling valid? - and I don't think valid? should make any changes to the db.