Add date range selector on signatures report
- add: date range selector (begin and end date)
- add: placeholder when no data is available for date range
- add: associated specs

- refactor: extract DateInput, responsible for parsing date input or
            returning a default value
- refactor: extract CSVReportGenerator, responsible for generating a CSV
            report given a number of records and property (method) names
- add: associated specs

This is a large commit. The primary success is that start and end dates
can now be selected on the signatures report page. While adding that
feature, I refactored some of the common logic in the
`SignaturesReportsController`, `SignaturesReport`, and `HoursReport` out
into separate objects to make for more maintainable shared code.

The date field currently acts as a text field; that will be fixed in a
future commit.