Fix intermittent failure on signature report page
The first test on the "admin views signature report" spec was failing
intermittently (~20-25% of the time); the first shift event–set for the
current date–was not showing up on the page.

This commit ensures that everything is using the application time zone,
sets the app time zone to Central (for NOLA use), sets the time
range in `SignaturesReport` to go through the end of the end date, and
refactors some of the spec support code.

All tests should now consistently pass.