Add removal buttons to work site show view
Work site show view now has a destroy button and an activate/deactivate
button. The buttons are all standard, blue buttons because I couldn't
figure out the built-in class for "alert" or "danger"–that can be added
later if we deem it important.

Activation/Deactivation now redirect back, which gives a more expected
user flow. I copied the `redirect_back` method from the Rails 5 source;
if we need it again, we can move it up to `Admin::ApplicationController` or
even `ApplicationController`.