add icons for various browsers
Resolves #88

The preferred blue-and-green-on-white design is used for the
public-facing form, while a black-on-white variant is used in the admin
dashboard to provide a visual cue as suggested in the issue.

I used rails_real_favicon to generate these icons from the EPS files
provided by the org, which I converted to SVG to serve as a "master"
icon.