Refactor non-administrate admin layouts
- remove: methods on ApplicationController
- add:    devise initializer for layout
- add:    admin layout to in-app admin controllers

On the plus side, this commit takes out the methods on
ApplicationController that customize only a small set of behavior. On
the minus side, this solution requires the layout location
(`'admin/non_administrate'`) to be listed three separate times, and
every additional admin controller will require another listing.

As it stands, putting this into a module or controller method won't save
much repetition and will add complexity. If more admin controllers are
added, we should think about encapsulating the layout name in a module
or method or something.