Add CONTRIBUTING.md
The goal of this file is to provide some basic contribution guidelines
for this project. It will make it easier for new contributors to get
started, and it will help us to ensure the quality of pull requests.

It was added as CONTRIBUTING.md on the project (as opposed to on the
wiki) because GitHub automatically links that file on the new PR page
(see https://github.com/blog/1184-contributing-guidelines); it also is
included in all forked/cloned projects for reference.