Rename a couple local variables
I think that round_trip might have meant something like “it was
originally like this, we saved it in the database and now we’ve loaded
it back again”. But I’m not really sure.

In my opinion, it is easier to understand these specs if we describe
what is in that variable - the rows. I can change it back if you guys
want.