Clean up `HoursReport#pull_join`
Created a `completed` scope for the shift model, and used it in
`HoursReport` so that we don’t grab unnecessary records from the
database.

In `completed`, the outer `from` is there to drop the table joined on
by `with_shift_end` so that later joining of the shift events table
(e.g. preloading with includes) doesn’t cause problems (I saw issues
with `report.pull_join.count` in the `HoursReport` spec).