Test form record creation in form spec
We should test the creation of the db records in a lower level test. One of the benefits of having a form object is that we can test in a unit test rather than a feature spec. It's nice when feature specs only test from the perspective of the end user.